Rename getCondition methods for clarity
Also added a few more. On build, webpack breaks
the links to the icons on the cdn.weatherapi.com
domain, so may have to download, import images
and implement manually. If you copy out the src
of the images and paste them into the browser,
they work fine, but in the web page they don't.