Fixed broken image links in dist version
For some reason, the links are fine in the dev
server just the way the weatherAPI provides them,
but in the build version all the images would display
broken links. You could copy the img src into your
browser and it would load fine but not in the page.
Turns out it works with 'http:' prepended to src.

Looking into protocol-relative URLs to try and make
sense of this.